Move Macro to parser package, now that macros live in the parser stage.
