+++ src/main/java/com/gnopai/ji65/SourceFileProcessor.java
@@ -1,6 +1,6 @@
 package com.gnopai.ji65;
 
-import com.gnopai.ji65.assembler.Macro;
+import com.gnopai.ji65.parser.Macro;
 import com.gnopai.ji65.parser.ParseletFactory;
 import com.gnopai.ji65.parser.Parser;
 import com.gnopai.ji65.parser.TokenStream;
