+++ src/main/java/com/gnopai/ji65/parser/statement/MacroCallParselet.java
@@ -1,7 +1,7 @@
 package com.gnopai.ji65.parser.statement;
 
 import com.gnopai.ji65.SourceFileProcessor;
-import com.gnopai.ji65.assembler.Macro;
+import com.gnopai.ji65.parser.Macro;
 import com.gnopai.ji65.parser.Parser;
 import com.gnopai.ji65.scanner.Token;
 import com.gnopai.ji65.scanner.TokenType;
