+++ src/test/java/com/gnopai/ji65/parser/statement/MacroCallParseletTest.java
@@ -2,7 +2,7 @@ package com.gnopai.ji65.parser.statement;
 
 import com.gnopai.ji65.InstructionType;
 import com.gnopai.ji65.SourceFileProcessor;
-import com.gnopai.ji65.assembler.Macro;
+import com.gnopai.ji65.parser.Macro;
 import com.gnopai.ji65.parser.ParseException;
 import com.gnopai.ji65.scanner.Token;
 import com.gnopai.ji65.util.ErrorHandler;
