+++ src/test/java/com/gnopai/ji65/parser/statement/MacroDirectiveParserTest.java
@@ -3,7 +3,7 @@ package com.gnopai.ji65.parser.statement;
 import com.gnopai.ji65.DirectiveType;
 import com.gnopai.ji65.InstructionType;
 import com.gnopai.ji65.SourceFileProcessor;
-import com.gnopai.ji65.assembler.Macro;
+import com.gnopai.ji65.parser.Macro;
 import com.gnopai.ji65.parser.ParseException;
 import com.gnopai.ji65.util.ErrorHandler;
 import org.junit.jupiter.api.Test;
